requirements.txt
yfinance
pandas
numpy
matplotlib
scikit-learn
joblib
fastapi 
uvicorn 
httpx 
python-dotenv
requests
supabase
